client-app/requirement.client.txt
Flask==3.1.0
requests==2.32.3
flask_cors==4.0.0
python-dotenv==1.0.1
